+++ HISTORY.md
@@ -12,6 +12,7 @@ and this project adheres to `Semantic Versioning: https://semver.org/spec/v2.0.0
 * HDLC client implementation
 * DlMS client implementation
 * Support for Get service including service specific block transfer
+* Support for selective access via range descriptor
 * Support for HLS authentication using HLS-GMAC.
 * Support for GlobalCiphering